Added async support
