+++ tests/collections/test_collections.py
@@ -17,60 +17,11 @@ Unit tests for the :mod:`pennylane.collection` submodule.
 from collections.abc import Sequence
 
 import pytest
-from pennylane import numpy as np
+import numpy as np
 
 import pennylane as qml
 
-
-try:
-    import torch
-except ImportError as e:
-    pass
-
-
-try:
-    import tensorflow as tf
-
-    if tf.__version__[0] == "1":
-        print(tf.__version__)
-        import tensorflow.contrib.eager as tfe
-
-        tf.enable_eager_execution()
-        Variable = tfe.Variable
-    else:
-        from tensorflow import Variable
-except ImportError as e:
-    pass
-
-
-@pytest.fixture
-def qnodes(interface, tf_support, torch_support):
-    """fixture returning some QNodes"""
-    if interface == "torch" and not torch_support:
-        pytest.skip("Skipped, no torch support")
-
-    if interface == "tf" and not tf_support:
-        pytest.skip("Skipped, no tf support")
-
-    dev1 = qml.device("default.qubit", wires=2)
-    dev2 = qml.device("default.qubit", wires=2)
-
-    @qml.qnode(dev1, interface=interface)
-    def qnode1(x):
-        qml.RX(x[0], wires=0)
-        qml.RY(x[1], wires=0)
-        qml.CNOT(wires=[0, 1])
-        return qml.var(qml.PauliZ(0)), qml.expval(qml.PauliZ(1))
-
-    @qml.qnode(dev2, interface=interface)
-    def qnode2(x):
-        qml.Hadamard(wires=0)
-        qml.RX(x[0], wires=0)
-        qml.RY(x[1], wires=1)
-        qml.CNOT(wires=[0, 1])
-        return qml.expval(qml.PauliZ(0)), qml.var(qml.PauliZ(1))
-
-    return qnode1, qnode2
+from conftest import torch, tf, Variable
 
 
 class TestMap:
@@ -426,240 +377,3 @@ class TestDot:
         """Test exception raised if no qnodes are provided as arguments"""
         with pytest.raises(ValueError, match="At least one argument must be a QNodeCollection"):
             qml.dot([1, 2], [3, 4])
-
-
-class TestQNodeCollection:
-    """Tests for the QNodeCollection class"""
-
-    def test_empty_init(self):
-        """Test that an empty QNode collection can be initialized"""
-        qc = qml.QNodeCollection()
-        assert qc.qnodes == []
-        assert len(qc) == 0
-
-    def test_init_with_qnodes(self):
-        """Test that a QNode collection can be initialized with QNodes"""
-        dev = qml.device("default.qubit", wires=1)
-
-        def circuit(x):
-            qml.RX(x, wires=0)
-            return qml.expval(qml.PauliZ(0))
-
-        qnodes = [qml.QNode(circuit, dev) for i in range(4)]
-        qc = qml.QNodeCollection(qnodes)
-
-        assert qc.qnodes == qnodes
-        assert len(qc) == 4
-
-    @pytest.mark.parametrize("interface", ["autograd", "numpy", "torch", "tf"])
-    def test_interface_property(self, interface, tf_support, torch_support):
-        """Test that the interface property correctly
-        resolves interfaces from the internal QNodes"""
-        if interface == "torch" and not torch_support:
-            pytest.skip("Skipped, no torch support")
-
-        if interface == "tf" and not tf_support:
-            pytest.skip("Skipped, no tf support")
-
-        qc = qml.QNodeCollection()
-        assert qc.interface is None
-
-        def circuit(x):
-            qml.RX(x, wires=0)
-            return qml.expval(qml.PauliZ(0))
-
-        dev = qml.device("default.qubit", wires=1)
-        qnodes = [qml.QNode(circuit, dev, interface=interface) for i in range(4)]
-        qc = qml.QNodeCollection(qnodes)
-
-        if interface == "numpy":
-            # Note: the "numpy" interface is deprecated, and
-            # now resolves to "autograd"
-            interface = "autograd"
-
-        assert qc.interface == interface
-
-    def test_append_qnode(self):
-        """Test that a QNode is correctly appended"""
-        qc = qml.QNodeCollection()
-        assert qc.qnodes == []
-
-        def circuit(x):
-            qml.RX(x, wires=0)
-            return qml.expval(qml.PauliZ(0))
-
-        dev = qml.device("default.qubit", wires=1)
-        qnode = qml.QNode(circuit, dev)
-        qc.append(qnode)
-
-        assert qc.qnodes == [qnode]
-
-    def test_extend_qnodes(self):
-        """Test that a list of QNodes is correctly appended"""
-        qc = qml.QNodeCollection()
-        assert qc.qnodes == []
-
-        def circuit(x):
-            qml.RX(x, wires=0)
-            return qml.expval(qml.PauliZ(0))
-
-        dev = qml.device("default.qubit", wires=1)
-        qnodes = [qml.QNode(circuit, dev) for i in range(4)]
-        qc.extend(qnodes)
-
-        assert qc.qnodes == [] + qnodes
-
-    def test_extend_multiple_interface_qnodes(self):
-        """Test that an error is returned if QNodes with differing
-        interfaces are attempted to be added to a QNodeCollection"""
-        qc = qml.QNodeCollection()
-
-        def circuit(x):
-            qml.RX(x, wires=0)
-            return qml.expval(qml.PauliZ(0))
-
-        dev = qml.device("default.qubit", wires=1)
-        qnodes = [
-            qml.QNode(circuit, dev, interface="autograd"),
-            qml.QNode(circuit, dev, interface=None),
-        ]
-
-        with pytest.raises(ValueError, match="do not all use the same interface"):
-            qc.extend(qnodes)
-
-    def test_extend_interface_mistmatch(self):
-        """Test that an error is returned if QNodes with a differing
-        interface to the QNode collection are appended"""
-        qc = qml.QNodeCollection()
-
-        def circuit(x):
-            qml.RX(x, wires=0)
-            return qml.expval(qml.PauliZ(0))
-
-        dev = qml.device("default.qubit", wires=1)
-        qnode1 = qml.QNode(circuit, dev, interface="autograd")
-        qnode2 = qml.QNode(circuit, dev, interface=None)
-
-        qc.extend([qnode1])
-
-        with pytest.raises(ValueError, match="Interface mismatch"):
-            qc.extend([qnode2])
-
-    def test_indexing(self):
-        """Test that indexing into the QNodeCollection correctly works"""
-
-        def circuit(x):
-            qml.RX(x, wires=0)
-            return qml.expval(qml.PauliZ(0))
-
-        dev = qml.device("default.qubit", wires=1)
-        qnodes = [qml.QNode(circuit, dev) for i in range(4)]
-
-        qc = qml.QNodeCollection(qnodes)
-        assert qc[2] == qnodes[2]
-
-    def test_sequence(self):
-        """Test that the QNodeCollection is a sequence type"""
-        qc = qml.QNodeCollection()
-        assert isinstance(qc, Sequence)
-
-    @pytest.mark.parametrize("interface", ["autograd", "numpy"])
-    def test_eval_autograd(self, qnodes):
-        """Test correct evaluation of the QNodeCollection using
-        the Autograd interface"""
-        qnode1, qnode2 = qnodes
-        qc = qml.QNodeCollection([qnode1, qnode2])
-        params = [0.5643, -0.45]
-
-        res = qc(params)
-        expected = np.vstack([qnode1(params), qnode2(params)])
-        assert np.all(res == expected)
-
-    @pytest.mark.parametrize("interface", ["autograd", "numpy"])
-    def test_grad_autograd(self, qnodes):
-        """Test correct gradient of the QNodeCollection using
-        the Autograd interface"""
-        qnode1, qnode2 = qnodes
-
-        params = [0.5643, -0.45]
-        qc = qml.QNodeCollection([qnode1, qnode2])
-        cost_qc = lambda params: np.sum(qc(params))
-        grad_qc = qml.grad(cost_qc, argnum=0)
-
-        cost_expected = lambda params: np.sum(qnode1(params) + qnode2(params))
-        grad_expected = qml.grad(cost_expected, argnum=0)
-
-        res = grad_qc(params)
-        expected = grad_expected(params)
-
-        assert np.all(res == expected)
-
-    @pytest.mark.parametrize("interface", ["torch"])
-    def test_eval_torch(self, qnodes, skip_if_no_torch_support):
-        """Test correct evaluation of the QNodeCollection using
-        the torch interface"""
-        qnode1, qnode2 = qnodes
-        qc = qml.QNodeCollection([qnode1, qnode2])
-        params = [0.5643, -0.45]
-
-        res = qc(params).numpy()
-        expected = np.vstack([qnode1(params), qnode2(params)])
-        assert np.all(res == expected)
-
-    @pytest.mark.parametrize("interface", ["torch"])
-    def test_grad_torch(self, qnodes, skip_if_no_torch_support):
-        """Test correct gradient of the QNodeCollection using
-        the torch interface"""
-        qnode1, qnode2 = qnodes
-
-        # calculate the gradient of the collection using pytorch
-        params = torch.autograd.Variable(torch.tensor([0.5643, -0.45]), requires_grad=True)
-        qc = qml.QNodeCollection([qnode1, qnode2])
-        cost = torch.sum(qc(params))
-        cost.backward()
-        res = params.grad.numpy()
-
-        # calculate the gradient of the QNodes individually using pytorch
-        params = torch.autograd.Variable(torch.tensor([0.5643, -0.45]), requires_grad=True)
-        cost = torch.sum(qnode1(params) + qnode2(params))
-        cost.backward()
-        expected = params.grad.numpy()
-
-        assert np.all(res == expected)
-
-    @pytest.mark.parametrize("interface", ["tf"])
-    def test_eval_tf(self, qnodes, skip_if_no_tf_support):
-        """Test correct evaluation of the QNodeCollection using
-        the tf interface"""
-        qnode1, qnode2 = qnodes
-        qc = qml.QNodeCollection([qnode1, qnode2])
-        params = [0.5643, -0.45]
-
-        res = qc(params).numpy()
-        expected = np.vstack([qnode1(params), qnode2(params)])
-        assert np.all(res == expected)
-
-    @pytest.mark.parametrize("interface", ["tf"])
-    def test_grad_tf(self, qnodes, skip_if_no_tf_support):
-        """Test correct gradient of the QNodeCollection using
-        the tf interface"""
-        qnode1, qnode2 = qnodes
-
-        # calculate the gradient of the collection using tf
-        params = Variable([0.5643, -0.45])
-        qc = qml.QNodeCollection([qnode1, qnode2])
-
-        with tf.GradientTape() as tape:
-            tape.watch(params)
-            cost = sum(qc(params))
-            res = tape.gradient(cost, params).numpy()
-
-        # calculate the gradient of the QNodes individually using tf
-        params = Variable([0.5643, -0.45])
-
-        with tf.GradientTape() as tape:
-            tape.watch(params)
-            cost = sum(qnode1(params) + qnode2(params))
-            expected = tape.gradient(cost, params).numpy()
-
-        assert np.all(res == expected)
